Rename networkstatsfetcher to networkstats, with a fetcher inside
