+++ src/code.cloudfoundry.org/netmon/network_stats/fetcher.go
@@ -1,28 +1,28 @@
-package network_stats_fetcher
+package network_stats
 
 import (
 	"code.cloudfoundry.org/lager"
 	"code.cloudfoundry.org/lib/rules"
 )
 
-//go:generate counterfeiter -o ../fakes/network_stats_fetcher.go --fake-name NetworkStatsFetcher . NetworkStatsFetcher
-type NetworkStatsFetcher interface {
+//go:generate counterfeiter -o ../fakes/network_stats_fetcher.go --fake-name NetworkStatsFetcher . Fetcher
+type Fetcher interface {
 	CountIPTablesRules() (int, error)
 }
 
-type networkStatsFetcher struct {
+type fetcher struct {
 	IPTablesAdapter rules.IPTablesAdapter
 	Logger          lager.Logger
 }
 
-func New(iptablesAdapter rules.IPTablesAdapter, logger lager.Logger) networkStatsFetcher {
-	return networkStatsFetcher{
+func NewFetcher(iptablesAdapter rules.IPTablesAdapter, logger lager.Logger) fetcher {
+	return fetcher{
 		IPTablesAdapter: iptablesAdapter,
 		Logger:          logger,
 	}
 }
 
-func (stats networkStatsFetcher) CountIPTablesRules() (int, error) {
+func (stats fetcher) CountIPTablesRules() (int, error) {
 	filterRules, err := stats.IPTablesAdapter.RuleCount("filter")
 	if err != nil {
 		stats.Logger.Error("failed-getting-filter-rules", err)
