+++ src/code.cloudfoundry.org/netmon/network_stats/network_stats_suite_test.go
@@ -1,0 +1,13 @@
+package network_stats_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestNetworkStats(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "NetworkStats Suite")
+}
