--- src/code.cloudfoundry.org/netmon/network_stats_fetcher/network_stats_fetcher_suite_test.go
@@ -1,13 +1,0 @@
-package network_stats_fetcher_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestNetworkStatsFetcher(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "NetworkStatsFetcher Suite")
-}
